requirements.txt
albumentations==0.5.2
matplotlib==3.2.1
networkx==2.3
numba==0.44.1
numpy==1.18.5
opencv_python==4.2.0.32
Pillow==9.3.0
ripser==0.6.4
scikit_image==0.19.1
scikit_learn==1.2.0
scipy==1.4.1
skimage==0.0
torch==1.8.0
torchvision==0.9.0
tqdm==4.32.1
visdom==0.1.8.9
